+++ service/verifiedpermissions/api_op_CreatePolicy.go
@@ -111,6 +111,15 @@ type CreatePolicyOutput struct {
 	// This member is required.
 	PolicyType types.PolicyType
 
+	// The action that a policy permits or forbids. For example, {"actions":
+	// [{"actionId": "ViewPhoto", "actionType": "PhotoFlash::Action"}, {"entityID":
+	// "SharePhoto", "entityType": "PhotoFlash::Action"}]} .
+	Actions []types.ActionIdentifier
+
+	// The effect of the decision that a policy returns to an authorization request.
+	// For example, "effect": "Permit" .
+	Effect types.PolicyEffect
+
 	// The principal specified in the new policy's scope. This response element isn't
 	// present when principal isn't specified in the policy content.
 	Principal *types.EntityIdentifier